+++ R/dct.R
@@ -19,6 +19,7 @@
 # Version history
 # 20201015  GvB       setup for gsignal v0.1.0
 # 20210506  GvB       use matrix() instead of as.matrix()
+# 20220328  GvB       copy dimnames of x to output object
 #------------------------------------------------------------------------------
 
 #' Discrete Cosine Transform
@@ -118,5 +119,6 @@
   if (vec) {
     y <- as.vector(y)
   }
+  dimnames(y) <- dimnames(x)
   y
 }