+++ R/decimate.R
@@ -18,6 +18,7 @@
 #
 # Version history
 # 20201129  GvB       setup for gsignal v0.1.0
+# 20220328  GvB       copy dimnames of x to output object
 #------------------------------------------------------------------------------
 
 #' Decrease sample rate
@@ -89,5 +90,6 @@ decimate <- function(x, q, n = ifelse(ftype == "iir", 8, 30), ftype = "iir") {
   if (vec) {
     y <- as.vector(y)
   }
+  dimnames(y) <- dimnames(x)
   y
 }